refactor(actions): split actions into dedicated module (attack/move/inspect)
