+++ backend/game/core/actions/action_inspect.py
@@ -3,60 +3,46 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-class MoveAction:
-    def __init__(
-        self,
-        state_manager,
-        runtime_player_service,
-        choice_service,
-        narrate_fn,
-        response_fn,
-        resolver=None,  
-    ):
+class InspectAction:
+    def __init__(self, state_manager, runtime_player_service, choice_service, narrate_fn, response_fn):
         self.state_manager = state_manager
         self.runtime_player_service = runtime_player_service
         self.choice_service = choice_service
         self.narrate_fn = narrate_fn
         self.response_fn = response_fn
-        self.resolver = resolver  
 
     def handle(self, parsed_input, world=None):
         room = parsed_input.get("room")
         user_id = parsed_input.get("user_id")
-        target = parsed_input.get("target")
 
         room_key = self.state_manager.normalize_room_id(room)
         room_obj = self.state_manager.get_or_create_room(room_key)
 
-        player = self.runtime_player_service.get_or_create(room_obj, user_id)
+        self.runtime_player_service.get_or_create(room_obj, user_id)
 
-        if not player:
-            return self.response_fn(
-                "move",
-                "Player not found",
-                {"error": "no_player"},
-            )
-
-        player.location = target or "unknown"
+        enemies = list(room_obj.enemies.keys())
 
         narration = self.narrate_fn(
-            "move",
-            {"location": player.location},
+            "inspect",
+            {
+                "room": room_key,
+                "enemies": enemies
+            },
             world,
         )
 
         choices = self.choice_service.build_choices(
             adventure_id=parsed_input.get("adventure"),
             room_key=room_key,
-            event_type="move",
-            result={"location": player.location},
+            event_type="inspect",
+            result={"room": room_key, "enemies": enemies},
             world=world,
         )
 
         return self.response_fn(
-            "move",
-            narration.get("text", f"Przemieszczasz się do: {player.location}"),
-            {"location": player.location},
+            "inspect",
+            narration.get("text", "Rozglądasz się po okolicy"),
+            {"room": room_key, "enemies": enemies},
             world,
             choices,
         )
